backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.2.1
beanie==2.0.0
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
fastapi==0.119.0
h11==0.16.0
httptools==0.7.1
idna==3.11
lazy-model==0.3.0
motor==3.7.1
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pymongo==4.15.3
PyJWT==2.9.0
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.1
websockets==15.0.1
